+++ web/documentserver-example/java-spring/src/main/resources/static/scripts/converter.js
@@ -25,9 +25,9 @@ var FillExtList;
 if (typeof jQuery !== "undefined") {
     jQuery.post('/config',
         function(data) {
-            FillExtList = data.FillExtList;
-            ConverExtList = data.ConverExtList;
-            EditedExtList = data.EditedExtList;
+            FillExtList = data.FillExtList.split(',');
+            ConverExtList = data.ConverExtList.split(',');
+            EditedExtList = data.EditedExtList.split(',');
             UrlConverter = data.UrlConverter;
             UrlEditor = data.UrlEditor;
     });