+++ web/documentserver-example/php/src/views/IndexView.php
@@ -63,8 +63,8 @@ final class IndexView extends View
     private function getLanguageListOptionsLayout()
     {
         $layout = "";
-        $config_manager = new ConfigurationManager();
-        foreach ($config_manager->languages() as $key => $language) {
+        $configManager = new ConfigurationManager();
+        foreach ($configManager->languages() as $key => $language) {
             $layout .= '<option value="'.$key.'">'.$language.'</option>'.PHP_EOL;
         }
         return $layout;
@@ -102,8 +102,8 @@ final class IndexView extends View
 
     private function getPreloaderUrl()
     {
-        $config_manager = new ConfigurationManager();
-        return $config_manager->document_server_preloader_url()->string();
+        $configManager = new ConfigurationManager();
+        return $configManager->documentServerPreloaderURL()->string();
     }
 
     private function getEditButton()