+++ CHANGELOG.md
@@ -1,19 +1,27 @@
 # Change Log
 
 - python: onRequestSelectSpreadsheet method
+- nodejs: link in referenceData
+- nodejs: onRequestSelectDocument method
 - nodejs: onRequestSelectSpreadsheet method
 - nodejs: onRequestOpen
 - nodejs: submitForm
 - nodejs: key in referenceData
 - nodejs: change reference source
+- java-spring: using a repo with a list of formats
+- java: using a repo with a list of formats
 - php: using a repo with a list of formats
 - nodejs: using a repo with a list of formats
+- java: using a repo with a list of formats
+- python: using a repo with a list of formats
+- ruby: using a repo with a list of formats
 - nodejs: delete file without reloading the page
 - nodejs: getting history by a separate request
-- nodejs: restore from history
-- python: restore from history
-- ruby: restore from history
-- php: restore from history
+- csharp-mvc: getting history by a separate request
+- csharp: getting history by a separate request
+- java: getting history by a separate request
+- java-spring: getting history by a separate request
+- restore from history
 
 ## 1.6.0
 - nodejs: setUsers for region protection